+++ src/Taskview/columns/index.ts
@@ -1,8 +1,5 @@
 export * from './AColumn';
-export * from './AttributeColumn';
 export { default as AttributeColumn } from './AttributeColumn';
 export * from './CohortColumn';
-export * from './NumberColumn';
 export { default as NumberColumn } from './NumberColumn';
-export * from './PrevalenceColumn';
 export { default as PrevalenceColumn } from './PrevalenceColumn';
