+++ src/Taskview/visualizations/GroupedBoxplot.ts
@@ -3,14 +3,14 @@ import log from 'loglevel';
 import { Spec as VegaSpec } from 'vega';
 import { TopLevelSpec as VegaLiteSpec } from 'vega-lite';
 import { ICohort } from '../../app/interfaces';
-import { IAttribute, IdValuePair } from '../../data/Attribute';
-import { NumRangeOperators } from '../../base/rest';
 import { IAttributeFilter, IFilterDesc } from '../../util';
 import { FilterEvent, SplitEvent } from '../../base/events';
 import { AVegaVisualization } from './AVegaVisualization';
 import { groupByConfig } from './config/GroupConfig';
 import { BRUSH_DATA_END, BRUSH_DATA_NAME, BRUSH_DATA_START, DATA_LABEL } from './constants';
 import { MultiAttributeVisualization } from './MultiAttributeVisualization';
+import { NumRangeOperators } from '../../base';
+import { IAttribute, IdValuePair } from '../../data';
 
 export class GroupedBoxplot extends MultiAttributeVisualization {
   static readonly NAME = 'Boxplot';
