+++ src/Taskview/visualizations/GroupedHistogram.ts
@@ -4,7 +4,6 @@ import { TopLevelUnitSpec } from 'vega-lite/build/src/spec/unit';
 import { getCohortLabel } from '../../Cohort';
 import { ICohort } from '../../app/interfaces';
 import { colors } from '../../config/colors';
-import { IdValuePair } from '../../data/Attribute';
 import { log } from '../../util';
 import { AVegaVisualization, FACETED_CHARTS_WIDTH } from './AVegaVisualization';
 import { groupByConfig } from './config/GroupConfig';
@@ -12,6 +11,7 @@ import { dataConfig } from './config/ScaleConfig';
 import { sortByConfig, sortOrderConfig } from './config/SortConfig';
 import { DATA_LABEL } from './constants';
 import { VegaHistogram } from './Histogram';
+import { IdValuePair } from '../../data/IAttribute';
 
 export class VegaGroupedHistogram extends VegaHistogram {
   static readonly COUNT = 'Count';
