+++ src/Taskview/visualizations/MultiAttributeVisualization.ts
@@ -3,9 +3,9 @@ import { format, select } from 'd3v7';
 import tippy from 'tippy.js';
 import vegaEmbed from 'vega-embed';
 import { TopLevelSpec as VegaLiteSpec } from 'vega-lite';
-import { Cohort, getCohortLabel } from '../../Cohort';
+import { getCohortLabel } from '../../Cohort';
 import { ICohort } from '../../app/interfaces';
-import { IAttribute, IdValuePair } from '../../data/Attribute';
+import { IAttribute, IdValuePair } from '../../data';
 import { log, selectLast } from '../../util';
 import { AVegaVisualization } from './AVegaVisualization';
 import { DATA_LABEL } from './constants';
