+++ src/Taskview/visualizations/Scatterplot.ts
@@ -5,8 +5,8 @@ import { TopLevelSpec as VegaLiteSpec } from 'vega-lite';
 import { TopLevel, LayerSpec } from 'vega-lite/build/src/spec';
 import { Field } from 'vega-lite/build/src/channeldef';
 import { ICohort } from '../../app/interfaces';
-import { AttributeType, IAttribute, IdValuePair, ServerColumnAttribute } from '../../data/Attribute';
-import { NumRangeOperators } from '../../base/rest';
+import { AttributeType, IAttribute, IdValuePair, ServerColumnAttribute } from '../../data';
+import { NumRangeOperators } from '../../base';
 import { IFilterDesc, inRange } from '../../util';
 import { FilterEvent } from '../../base/events';
 import { DATA_LABEL } from './constants';
