+++ src/cohortview.ts
@@ -1,6 +1,6 @@
 import { IObjectRef, ProvenanceGraph, IDatabaseViewDesc, RestBaseUtils } from 'tdp_core';
 import { CohortSelectionListener } from './app/CoralSelectionListener';
-import { CoralApp } from './app/CoralApp';
+import type { CoralApp } from './app/CoralApp';
 import { OnboardingManager } from './OnboardingManager';
 import { CohortOverview } from './Overview/CohortOverview';
 import { RectangleLayout } from './Overview/OverviewLayout';
@@ -8,11 +8,31 @@ import Taskview from './Taskview/Taskview';
 import { handleDataLoadError, log } from './util';
 import { IEntitySourceConfig } from './config/entities';
 import { ICohort } from './app/interfaces';
+import { CohortContext } from './CohortContext';
 
-export let cohortOverview: CohortOverview;
-export let taskview: Taskview;
+export function destroyOld() {
+  if (CohortContext.taskview) {
+    CohortContext.taskview.destroy();
+    CohortContext.taskview = null;
+  }
+  if (CohortContext.cohortOverview) {
+    CohortContext.cohortOverview.destroy();
+    CohortContext.cohortOverview = null;
+  }
+  CohortContext.referenceCohort = null;
+}
 
-let referenceCohort: ICohort;
+export async function loadViewDescription(database: string, view: string) {
+  log.debug('getTDPDesc for: db:', database, ' |view: ', view);
+  try {
+    const descr: IDatabaseViewDesc = await RestBaseUtils.getTDPDesc(database, view);
+    log.debug('descr= ', descr);
+    return descr;
+  } catch (e) {
+    handleDataLoadError(e);
+    return null;
+  }
+}
 
 export async function createCohortOverview(
   graph: ProvenanceGraph,
@@ -28,60 +48,21 @@ export async function createCohortOverview(
   destroyOld();
 
   // set reference/root cohort
-  referenceCohort = rootCohort; // save reference cohort
+  CohortContext.referenceCohort = rootCohort; // save reference cohort
   log.debug('cohortview - root: ', rootCohort);
 
   // create Overview
-  cohortOverview = new CohortOverview(container, graph, ref, new RectangleLayout(130, 50, 50, 50), rootCohort, viewDescription);
+  CohortContext.cohortOverview = new CohortOverview(container, graph, ref, new RectangleLayout(130, 50, 50, 50), rootCohort, viewDescription);
   // create Taskview
-  taskview = new Taskview(detailView, rootCohort);
+  CohortContext.taskview = new Taskview(detailView, rootCohort);
 
-  CohortSelectionListener.get().taskview = taskview;
+  CohortSelectionListener.get().taskview = CohortContext.taskview;
   // draw overview in container element
-  updateOverview(cohortOverview);
+  CohortContext.cohortOverview.generateOverview();
 
   OnboardingManager.addTip('rootCohort', rootCohort.representation.getRepresentation());
   return {
-    cohortOV: cohortOverview,
-    taskV: taskview,
+    cohortOV: CohortContext.cohortOverview,
+    taskV: CohortContext.taskview,
   };
-}
-
-export function getRootCohort(): ICohort {
-  return referenceCohort;
-}
-
-export function destroyOld() {
-  if (taskview) {
-    taskview.destroy();
-    taskview = null;
-  }
-  if (cohortOverview) {
-    cohortOverview.destroy();
-    cohortOverview = null;
-  }
-  referenceCohort = null;
-}
-
-export function getCohortOverview(): CohortOverview {
-  return cohortOverview;
-}
-
-export function getTaskview(): Taskview {
-  return taskview;
-}
-
-export function updateOverview(overview: CohortOverview) {
-  overview.generateOverview();
-}
-
-export async function loadViewDescription(database: string, view: string) {
-  log.debug('getTDPDesc for: db:', database, ' |view: ', view);
-  try {
-    const descr: IDatabaseViewDesc = await RestBaseUtils.getTDPDesc(database, view);
-    log.debug('descr= ', descr);
-    return descr;
-  } catch (e) {
-    handleDataLoadError(e);
-  }
 }
