+++ src/data/IAttribute.ts
