+++ src/data/index.ts
@@ -1,1 +1,2 @@
 export * from './Attribute';
+export * from './IAttribute';
